refactored MessageLogic to accept any IMessagingService

+++ src/main/java/at/halora/services/IMessagingService.java
@@ -1,5 +1,5 @@
 package at.halora.services;
 
 public interface IMessagingService {
-    void sendMessage(Integer id, String message);
+    void sendMessage(Long id, String message);
 }
